Further transition to bent and simpler await / Promise structure
In addition starting to change the behaviour of the library and
parameter names to make entities a bit easier to identify by name.

Bigger change is it remove the option to use '_' as a placeholder
for accessing base images from e.g. Docker hub. Still supporting
this feature would harm the attempt to have a very thin library as
it would
  1st: hide Docker hub specific implementation details
  2nd: make is less obvious what happens in the background for code
       reviews by not-involved developers.

The test-tags.js file is just a temporary test file until we have
proper test infrastructure.